+++ src/main/java/com/eddi/repository/ReportRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReportRepository extends CrudRepository<Report, Integer> {
+public interface ReportRepository extends CrudRepository<Report, Long> {
 
     @Query(value = "SELECT * FROM report WHERE id =?1",
             nativeQuery = true)
-    Report findReportById(Integer id);
+    Report findReportById(Long id);
 }
